refactor: search moved inside post API
