--- src/search/search.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PrismaModule } from 'src/prisma/prisma.module';
-import { SearchRepository } from './search.repository';
-import { SearchController } from './search.controller';
-import { SearchService } from './search.service';
-
-@Module({
-  imports: [PrismaModule],
-  controllers: [SearchController],
-  providers: [SearchRepository, SearchService],
-})
-export class SearchModule {}
